pyproject.toml
[tool.poetry]
name = "desk-square"
version = "0.1.0"
description = "API for an event management tool (ex. Meetup and EvenBrite) written in drf django."
authors = ["Temiloluwa Samuel <65780579+stemitom@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
black = "^22.3.0"
djangorestframework-simplejwt = "^5.1.0"
flake8 = "^4.0.1"
coverage = "^6.3.2"
setuptools = "^62.3.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
